ai-fashion-demo/ai-fashion-demo/backend/requirements.txt
requests
flask
flask-cors
python-dotenv
huggingface_hub
pillow
